Refactor navigation menu component and update imports for consistency
