+++ .env
@@ -2,3 +2,4 @@ NEXT_PUBLIC_CSC_API_KEY=SThkSGZBV3Z4amdiSVduRlp0SkE4MEpwMnU4UWhpM2xOdDJERE5uWA==
 NEXT_PUBLIC_BASE_URL_USERS="https://api.inmero.co/sigma/users"
 NEXT_PUBLIC_BASE_URL_MAIN="https://api.inmero.co/sigma/main"
 NEXT_PUBLIC_BASE_URL_LOCATION="https://api.countrystatecity.in/v1"
+NEXT_PUBLIC_BASE_URL_WS="https://api.inmero.co/sigma/ws"
