+++ src/app/(dashboard)/parametrization/styles/page.jsx
@@ -1,7 +1,7 @@
 "use client";
 import React, { useState, useEffect } from 'react';
 import { FiEdit3, FiBell, FiPlus, FiX } from 'react-icons/fi';
-import NavigationMenu from '../../../components/parametrization/ParameterNavigation';
+import NavigationMenu from '../../../components/shared/ParameterNavigation';
 import ColorPickerModal from '../../../components/parametrization/ColorPickerModal';
 import NewThemeModal from '../../../components/parametrization/NewThemeModal';
 import { useTheme } from '@/contexts/ThemeContext';
